moved special logic to map-utils file
