+++ src/app/features/location/map/map.component.ts
@@ -10,22 +10,7 @@ import { Observable, Subject } from "rxjs";
 import { debounceTime, map } from "rxjs/operators";
 import { Coordinates } from "../coordinates";
 import { Entity } from "../../../core/entity/model/entity";
-import { getHueForEntity } from "../../../utils/utils";
-
-const iconRetinaUrl = "assets/marker-icon-2x.png";
-const iconUrl = "assets/marker-icon.png";
-const shadowUrl = "assets/marker-shadow.png";
-const iconDefault = L.icon({
-  iconRetinaUrl,
-  iconUrl,
-  shadowUrl,
-  iconSize: [25, 41],
-  iconAnchor: [12, 41],
-  popupAnchor: [1, -34],
-  tooltipAnchor: [16, -28],
-  shadowSize: [41, 41],
-});
-L.Marker.prototype.options.icon = iconDefault;
+import { getHueForEntity } from "../map-utils";
 
 @Component({
   selector: "app-map",
